Add new package: server
This package is used to orchestrate everything. It prints the initial
start text, creates the channels used between the incoming
protocol and renderers, and starts the goroutines.

Server has one protocol, which we use for incoming requests. At
this point it's only http.

Sever can have many renderers, which we use to output the
incoming requests from the protocol. At this point we have two,
printing to the terminal, or writing to a log file.