Add error channel, remove protocol from wg
This is use when a renderer fails to start or encounters an error.

When the above happens, the renderer will send an error signal
to the protocol server, which will tell the other renderers to quit,
and eventually close the waitgroup.

Also removed the protocol from the wg because we can assume
that if all renderers have exited, we should end the main program
because there is nothing to output to.