Rename Params to Message
We pass the params as a struct and this field is mainly used
to print a string, which is more useful if it is a generic message.

We plan to use this to send websocket messages as well as any
other string type data we want to send to the printers.